refactor: Создала компонент ActionItem
