+++ src/entities/ActionItem/index.ts
@@ -1,0 +1,1 @@
+export { ActionItem } from "./ui";
